pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ml-platform"
version = "0.1.0"
description = "Shared infrastructure for production ML services: serving, monitoring, tracking, and deployment."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Annette Taberner-Miller" }]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
stateful = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "boto3>=1.34",
]
stateless = [
    "bentoml>=1.4",
]
monitoring = [
    "boto3>=1.34",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
tracking = [
    "mlflow>=2.10",
    "boto3>=1.34",
]
infra = [
    "aws-cdk-lib>=2.130",
    "constructs>=10.0",
]
auth = [
    "PyJWT>=2.8",
    "fastapi>=0.110",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "asgi-lifespan>=2.1",
    "moto[cloudformation,dynamodb,ecr,ecs,elbv2,s3,ses,sqs,secretsmanager,sts]>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
    "PyJWT>=2.8",
]
all = [
    "ml-platform[stateful,stateless,monitoring,tracking,infra,auth,dev]",
]

[project.scripts]
ml-platform = "ml_platform.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_platform"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
